refactor: RefreshTokenEntity를 core 모듈의 domain.entity로 이동
- 기존 auth 모듈 내부에 있던 RefreshTokenEntity를 core.domain.entity로 분리
- 공통 도메인 엔티티로 분리하여 다른 모듈(admin 등)에서도 재사용 가능하도록 구조 개선
- 관련 JPA 레포지토리 및 서비스에서 import 경로 수정
