+++ auth/src/main/java/kernel360/ckt/auth/infra/RefreshTokenJpaRepository.java
@@ -1,5 +1,6 @@
 package kernel360.ckt.auth.infra;
 
+import kernel360.ckt.core.domain.entity.RefreshTokenEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.Optional;
